Updating states, file name, and documentation
